pyproject.toml
[project]
name = "hyper-leezus"
version = "0.1.0"
description = "Distributed sports analytics and prediction platform"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<0.120.0",
  "uvicorn[standard]>=0.32.0,<0.35.0",
  "pydantic>=2.9.0,<3.0.0",
  "pydantic-settings>=2.6.0,<3.0.0",
  "httpx>=0.28.0,<0.30.0",
  "pandas>=2.2.3,<2.3.0",
  "polars>=1.14.0,<1.15.0",
  "dask[dataframe,bag]>=2024.10.0,<2025.0.0",
  "sqlalchemy>=2.0.36,<2.1.0",
  "psycopg[binary]>=3.2.3,<4.0.0",
  "redis>=4.2.2,<5.0.0",
  "boto3>=1.35.0,<1.36.0",
  "kafka-python>=2.0.2,<3.0.0",
  "pyarrow>=18.0.0,<19.0.0",
  "xgboost>=2.1.2,<2.2.0",
  "lightgbm>=4.5.0,<4.6.0",
  "scikit-learn>=1.5.2,<1.6.0",
  "torch>=2.5.1,<2.6.0",
  "pytorch-lightning>=2.4.0,<2.5.0",
  "optuna>=4.1.0,<4.2.0",
  "mlflow>=2.18.0,<2.19.0",
  "feast[redis,postgres]>=0.42.0,<0.43.0",
  "shap>=0.46.0,<0.47.0",
  "networkx>=3.4.2,<4.0.0",
  "pyjwt>=2.9.0,<3.0.0",
  "slowapi>=0.1.9,<0.2.0",
  "structlog>=24.4.0,<25.0.0",

]

[tool.setuptools]
py-modules = []

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
pythonpath = ["."]
